chore(queue): don't import actions in queue module
This will help to avoid circular module dependencies.
